enviroment stuff and refactor of placement

+++ lib/providers/pairing_provider.dart
@@ -3,8 +3,12 @@ import 'package:flutter/material.dart';
 import 'package:jhopping_list/db/database.dart';
 
 class PairingProvider extends ChangeNotifier {
-  Future<void> addHttpServerToRemoteTerminal(String terminalId, String host, int port, String nick) async {
+  Future<void> addHttpServerToRemoteTerminal(String terminalId, String host, int port, String nick, String enviromentId) async {
     final database = AppDatabaseSingleton.instance;
+
+    await database
+        .into(database.remoteTerminalEnviroments)
+        .insert(RemoteTerminalEnviromentsCompanion(terminalId: Value(terminalId), enviromentId: Value(enviromentId)));
 
     await database
         .into(database.remoteTerminals)
@@ -21,8 +25,12 @@ class PairingProvider extends ChangeNotifier {
     notifyListeners();
   }
 
-  Future<void> addHttpClientToRemoteTerminal(String terminalId, String nick) async {
+  Future<void> addHttpClientToRemoteTerminal(String terminalId, String nick, String enviromentId) async {
     final database = AppDatabaseSingleton.instance;
+
+    await database
+        .into(database.remoteTerminalEnviroments)
+        .insert(RemoteTerminalEnviromentsCompanion(terminalId: Value(terminalId), enviromentId: Value(enviromentId)));
 
     await database
         .into(database.remoteTerminals)
@@ -31,24 +39,30 @@ class PairingProvider extends ChangeNotifier {
     notifyListeners();
   }
 
-  Future<void> updateNick(String terminalId, String nick) async {
+  Future<List<RemoteTerminal>> getRemoteTerminals(String enviromentId) async {
     final database = AppDatabaseSingleton.instance;
 
-    await (database.update(database.remoteTerminals)
-      ..where((table) => table.terminalId.equals(terminalId))).write(RemoteTerminalsCompanion(nick: Value(nick)));
-
-    notifyListeners();
+    return await (database.select(database.remoteTerminals)
+          ..join([
+            innerJoin(
+              database.remoteTerminalEnviroments,
+              database.remoteTerminalEnviroments.terminalId.equalsExp(database.remoteTerminals.terminalId),
+            ),
+          ])
+          ..where((tbl) => database.remoteTerminalEnviroments.enviromentId.equals(enviromentId)))
+        .get();
   }
 
-  Future<List<RemoteTerminal>> getRemoteTerminals() async {
+  Future<void> deleteRemoteTerminalById(String terminalId, String enviromentId) async {
     final database = AppDatabaseSingleton.instance;
 
-    return await database.select(database.remoteTerminals).get();
-  }
+    await (database.delete(database.remoteTerminalEnviroments)
+      ..where((table) => table.terminalId.equals(terminalId) & table.enviromentId.equals(enviromentId))).go();
 
-  Future<void> deleteRemoteTerminalById(String terminalId) async {
-    final database = AppDatabaseSingleton.instance;
-    await (database.delete(database.remoteTerminals)..where((table) => table.terminalId.equals(terminalId))).go();
+    if ((await (database.select(database.remoteTerminalEnviroments)..where((table) => table.terminalId.equals(terminalId))).get()).isEmpty) {
+      await (database.delete(database.remoteTerminals)..where((table) => table.terminalId.equals(terminalId))).go();
+    }
+
     notifyListeners();
   }
